refactor: restructure to subdomain-based package layout
- Domain: wallet/, remittance/, claim/, fx/, common/ (each with model/port/exception)
- Infrastructure: db/wallet/, db/remittance/, db/claim/ (replaces flat persistence/)
- Tests: mirror domain subdomain structure
- Delete inbound port interfaces (handlers will replace in feature branches)
- All ArchUnit rules pass with new structure
