+++ backend/src/main/java/com/stablepay/domain/common/port/SmsProvider.java
@@ -1,0 +1,5 @@
+package com.stablepay.domain.common.port;
+
+public interface SmsProvider {
+    void sendSms(String phoneNumber, String message);
+}
