--- backend/src/main/java/com/stablepay/domain/exception/InsufficientBalanceException.java
@@ -1,15 +1,0 @@
-package com.stablepay.domain.exception;
-
-import java.math.BigDecimal;
-
-public class InsufficientBalanceException extends RuntimeException {
-
-    public static InsufficientBalanceException forAmount(BigDecimal requested, BigDecimal available) {
-        return new InsufficientBalanceException(
-                "SP-0002: Insufficient balance. Requested: " + requested + ", Available: " + available);
-    }
-
-    private InsufficientBalanceException(String message) {
-        super(message);
-    }
-}
