+++ backend/src/main/java/com/stablepay/domain/fx/port/FxRateProvider.java
@@ -1,0 +1,7 @@
+package com.stablepay.domain.fx.port;
+
+import com.stablepay.domain.fx.model.FxQuote;
+
+public interface FxRateProvider {
+    FxQuote getRate(String fromCurrency, String toCurrency);
+}
