--- backend/src/main/java/com/stablepay/domain/model/RemittanceStatus.java
@@ -1,10 +1,0 @@
-package com.stablepay.domain.model;
-
-public enum RemittanceStatus {
-    INITIATED,
-    ESCROWED,
-    CLAIMED,
-    DELIVERED,
-    REFUNDED,
-    CANCELLED
-}
