--- backend/src/main/java/com/stablepay/domain/port/inbound/ClaimService.java
@@ -1,10 +1,0 @@
-package com.stablepay.domain.port.inbound;
-
-import java.util.Optional;
-
-import com.stablepay.domain.model.Remittance;
-
-public interface ClaimService {
-    Optional<Remittance> findByClaimToken(String token);
-    Remittance claim(String token, String upiId);
-}
