--- backend/src/main/java/com/stablepay/domain/port/outbound/ClaimTokenRepository.java
@@ -1,10 +1,0 @@
-package com.stablepay.domain.port.outbound;
-
-import java.util.Optional;
-
-import com.stablepay.domain.model.ClaimToken;
-
-public interface ClaimTokenRepository {
-    ClaimToken save(ClaimToken claimToken);
-    Optional<ClaimToken> findByToken(String token);
-}
