--- backend/src/main/java/com/stablepay/domain/port/outbound/FxRateProvider.java
@@ -1,7 +1,0 @@
-package com.stablepay.domain.port.outbound;
-
-import com.stablepay.domain.model.FxQuote;
-
-public interface FxRateProvider {
-    FxQuote getRate(String fromCurrency, String toCurrency);
-}
