--- backend/src/main/java/com/stablepay/domain/port/outbound/MpcWalletClient.java
@@ -1,6 +1,0 @@
-package com.stablepay.domain.port.outbound;
-
-public interface MpcWalletClient {
-    String generateKey();
-    byte[] signTransaction(byte[] transactionBytes, byte[] keyShareData);
-}
