--- backend/src/main/java/com/stablepay/domain/port/outbound/SmsProvider.java
@@ -1,5 +1,0 @@
-package com.stablepay.domain.port.outbound;
-
-public interface SmsProvider {
-    void sendSms(String phoneNumber, String message);
-}
