+++ backend/src/main/java/com/stablepay/domain/remittance/model/RemittanceStatus.java
@@ -1,0 +1,10 @@
+package com.stablepay.domain.remittance.model;
+
+public enum RemittanceStatus {
+    INITIATED,
+    ESCROWED,
+    CLAIMED,
+    DELIVERED,
+    REFUNDED,
+    CANCELLED
+}
