+++ backend/src/main/java/com/stablepay/domain/wallet/exception/InsufficientBalanceException.java
@@ -1,0 +1,15 @@
+package com.stablepay.domain.wallet.exception;
+
+import java.math.BigDecimal;
+
+public class InsufficientBalanceException extends RuntimeException {
+
+    public static InsufficientBalanceException forAmount(BigDecimal requested, BigDecimal available) {
+        return new InsufficientBalanceException(
+                "SP-0002: Insufficient balance. Requested: " + requested + ", Available: " + available);
+    }
+
+    private InsufficientBalanceException(String message) {
+        super(message);
+    }
+}
