+++ backend/src/main/java/com/stablepay/domain/wallet/port/MpcWalletClient.java
@@ -1,0 +1,6 @@
+package com.stablepay.domain.wallet.port;
+
+public interface MpcWalletClient {
+    String generateKey();
+    byte[] signTransaction(byte[] transactionBytes, byte[] keyShareData);
+}
