+++ backend/src/main/java/com/stablepay/domain/wallet/port/WalletRepository.java
@@ -1,0 +1,11 @@
+package com.stablepay.domain.wallet.port;
+
+import java.util.Optional;
+
+import com.stablepay.domain.wallet.model.Wallet;
+
+public interface WalletRepository {
+    Wallet save(Wallet wallet);
+    Optional<Wallet> findById(Long id);
+    Optional<Wallet> findByUserId(String userId);
+}
