+++ backend/src/main/java/com/stablepay/infrastructure/db/claim/ClaimTokenEntityMapper.java
@@ -1,0 +1,11 @@
+package com.stablepay.infrastructure.db.claim;
+
+import org.mapstruct.Mapper;
+
+import com.stablepay.domain.claim.model.ClaimToken;
+
+@Mapper(componentModel = "spring")
+public interface ClaimTokenEntityMapper {
+    ClaimToken toDomain(ClaimTokenEntity entity);
+    ClaimTokenEntity toEntity(ClaimToken domain);
+}
