+++ backend/src/main/java/com/stablepay/infrastructure/db/claim/ClaimTokenJpaRepository.java
@@ -1,0 +1,9 @@
+package com.stablepay.infrastructure.db.claim;
+
+import java.util.Optional;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+interface ClaimTokenJpaRepository extends JpaRepository<ClaimTokenEntity, Long> {
+    Optional<ClaimTokenEntity> findByToken(String token);
+}
