+++ backend/src/main/java/com/stablepay/infrastructure/db/remittance/RemittanceJpaRepository.java
@@ -1,0 +1,12 @@
+package com.stablepay.infrastructure.db.remittance;
+
+import java.util.List;
+import java.util.Optional;
+import java.util.UUID;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+interface RemittanceJpaRepository extends JpaRepository<RemittanceEntity, Long> {
+    Optional<RemittanceEntity> findByRemittanceId(UUID remittanceId);
+    List<RemittanceEntity> findBySenderId(String senderId);
+}
