--- backend/src/main/java/com/stablepay/infrastructure/persistence/ClaimTokenEntity.java
@@ -1,47 +1,0 @@
-package com.stablepay.infrastructure.persistence;
-
-import java.time.Instant;
-import java.util.UUID;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Entity
-@Table(name = "claim_tokens")
-@Getter
-@Setter
-@Builder(toBuilder = true)
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-public class ClaimTokenEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(name = "remittance_id", nullable = false)
-    private UUID remittanceId;
-
-    @Column(name = "token", nullable = false, unique = true)
-    private String token;
-
-    @Column(name = "claimed", nullable = false)
-    private boolean claimed;
-
-    @Column(name = "created_at", nullable = false, updatable = false)
-    private Instant createdAt;
-
-    @Column(name = "expires_at")
-    private Instant expiresAt;
-}
