--- backend/src/main/java/com/stablepay/infrastructure/persistence/ClaimTokenJpaRepository.java
@@ -1,9 +1,0 @@
-package com.stablepay.infrastructure.persistence;
-
-import java.util.Optional;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-interface ClaimTokenJpaRepository extends JpaRepository<ClaimTokenEntity, Long> {
-    Optional<ClaimTokenEntity> findByToken(String token);
-}
