--- backend/src/main/java/com/stablepay/infrastructure/persistence/WalletEntityMapper.java
@@ -1,11 +1,0 @@
-package com.stablepay.infrastructure.persistence;
-
-import org.mapstruct.Mapper;
-
-import com.stablepay.domain.model.Wallet;
-
-@Mapper(componentModel = "spring")
-public interface WalletEntityMapper {
-    Wallet toDomain(WalletEntity entity);
-    WalletEntity toEntity(Wallet domain);
-}
